+++ src/Domain/State/Phases/Phase.cs
@@ -16,7 +16,7 @@ namespace EnduranceJudge.Domain.State.Phases
 
         public bool IsFinal { get; internal set; }
         public int OrderBy { get; internal set; }
-        public int LengthInKm { get; internal set; }
+        public double LengthInKm { get; internal set; }
         public int MaxRecoveryTimeInMins { get; internal set; }
         public int RestTimeInMins { get; internal set; }
     }
